sensor-ui/requirements.txt
bottle==0.11.6
pyusb>=1.0.0b2
